refactor: 사용하지 않는 라이브러리와 hook 삭제, 무분별한 constants 모듈화 통합
