--- constants/animation.ts
@@ -1,6 +1,0 @@
-export const HeaderAnimation = {
-  ready: {
-    opacity: 0,
-  },
-  start: { opacity: 1, transition: { duration: 1, delay: 0.1 } },
-} as const;
