+++ constants/index.ts
@@ -9,7 +9,4 @@ export * from './application';
 export * from './schedule';
 export * from './net';
 export * from './message';
-export * from './login';
 export * from './common';
-export * from './promotion';
-export * from './animation';
