--- constants/login.ts
@@ -1,4 +1,0 @@
-export const LOGIN_DEFAULT_VALUE = {
-  id: '',
-  password: '',
-} as const;
