--- constants/promotion.ts
@@ -1,5 +1,0 @@
-export const PROMOTION_TYPE = {
-  PROJECT: 'Project',
-  STUDY: 'Study & Networking',
-  ENTERTAINMENT: 'Entertainment',
-} as const;
