+++ hooks/index.ts
@@ -7,7 +7,6 @@ export { default as useInputRef } from './useInputRef';
 export { default as useDeviceChecker } from './useDeviceChecker';
 export { default as useBeforeUnload } from './useBeforeUnload';
 export { default as useMousePosition } from './useMousePosition';
-export { default as useIsLoading } from './useIsLoading';
 export { default as useViewport } from './useViewport';
 export { default as useApplyContext } from './useApplyContext';
 export { default as useApplicationExplore } from './useApplicationExplore';
