--- hooks/useIsLoading.ts
@@ -1,14 +1,0 @@
-'use client';
-
-import { useEffect, useState } from 'react';
-
-const useIsLoading = () => {
-  const [isLoading, setIsLoading] = useState<boolean>(true);
-
-  useEffect(() => {
-    setIsLoading(false);
-  });
-  return isLoading;
-};
-
-export default useIsLoading;
