+++ package.json
@@ -12,11 +12,9 @@
     "next": "^14.1.0",
     "next-auth": "^4.24.6",
     "react": "^18.2.0",
-    "react-device-detect": "^2.2.3",
     "react-dom": "^18.2.0",
     "styled-components": "^6.0.7",
-    "sweetalert2": "^11.7.22",
-    "web-vitals": "^2.1.4"
+    "sweetalert2": "^11.7.22"
   },
   "scripts": {
     "dev": "next dev",
